Reformatting

+++ src/main/java/io/github/fukkitmc/fukkit/redirects/AbstractDecorationEntityRedirects.java
@@ -28,9 +28,9 @@ public interface AbstractDecorationEntityRedirects {
             if (entity != null) {
                 entity.setPos(d0, d1, d2);
             }
-            double d6 = (double) width;
-            double d7 = (double) height;
-            double d8 = (double) width;
+            double d6 = width;
+            double d7 = height;
+            double d8 = width;
 
             if (direction.getAxis() == Direction.Axis.Z) {
                 d8 = 1.0D;
